+++ src/main/java/com/driver/repository/AdminRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 import com.driver.model.Admin;
 //@Repository
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
 }
