+++ src/main/java/com/driver/repository/CabRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.driver.model.Cab;
 
 @Repository
-public interface CabRepository extends JpaRepository<Cab, Integer> {
+public interface CabRepository extends JpaRepository<Cab, Long> {
 
 }
