+++ src/main/java/com/driver/repository/DriverRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import com.driver.model.Driver;
 @Repository
-public interface DriverRepository extends JpaRepository<Driver, Integer>{
+public interface DriverRepository extends JpaRepository<Driver, Long>{
 
     List<Driver> findByAvailableTrueOrderByDriverIdAsc();
 }
