+++ src/main/java/com/driver/services/DriverService.java
@@ -5,6 +5,6 @@ import com.driver.model.Driver;
 public interface DriverService {
 
 		public void register(String mobile, String password);
-		public void removeDriver(int driverId);
-		public void updateStatus(int driverId);
+		public void removeDriver(Long driverId);
+		public void updateStatus(Long driverId);
 }
